Adding timing to Hadoop and reimplementations for spark

+++ hadoop-terasort/scripts/single-hadoop-setup.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
-export JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64
-export HADOOP_PREFIX=$(pwd)/download/hadoop-2.7.4
+master="flamedragon"
 
-master=$(head -n 1 hadoop-config.cfg)
+if [ ! $(dpkg -s openjdk-8-jdk >> /dev/null) ]; then
+sudo apt update
+sudo apt install -y openjdk-8-jdk
+fi
+export JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64
+export HADOOP_HOME=$(pwd)/download/hadoop-2.7.4
 
 if [ ! -d "download" ]
 then
@@ -24,28 +28,29 @@ then
     cd ..
 fi
 
-sudo apt update
-sudo apt install -y openjdk-8-jdk
-
 exists=$(cat ~/.bashrc | grep "JAVA_HOME" | wc -l)
-if [ $exists -eq 0 ]
+if [ ${exists} -eq 0 ]
 then
     echo "export JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64" >> ~/.bashrc
 fi
 
-exists=$(cat ~/.bashrc | grep "HADOOP_PREFIX" | wc -l)
-if [ $exists -eq 0 ]
+exists=$(cat ~/.bashrc | grep "HADOOP_HOME" | wc -l)
+if [ ${exists} -eq 0 ]
 then
-    echo "export HADOOP_PREFIX=$(pwd)/download/hadoop-2.7.4" >> ~/.bashrc
+    echo "export HADOOP_HOME=$(pwd)/download/hadoop-2.7.4" >> ~/.bashrc
+    echo "export LD_LIBRARY_PATH=$HADOOP_HOME/lib/native:$LD_LIBRARY_PATH" >> ~/.bashrc
+    echo "export PATH=${HADOOP_HOME}/bin:${HADOOP_HOME}/sbin:$PATH" >> ~/.bashrc
 fi
 
-head -n 24 $HADOOP_PREFIX/etc/hadoop/hadoop-env.sh > temp.txt
+head -n 24 ${HADOOP_HOME}/etc/hadoop/hadoop-env.sh > temp.txt
 echo "export JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64" >> temp.txt
-echo "export HADOOP_PREFIX=$(pwd)/download/hadoop-2.7.4" >> temp.txt
-head -n -25 $HADOOP_PREFIX/etc/hadoop/hadoop-env.sh >> temp.txt
-cat temp.txt > $HADOOP_PREFIX/etc/hadoop/hadoop-env.sh
+echo "export HADOOP_HOME=$(pwd)/download/hadoop-2.7.4" >> temp.txt
+head -n -25 ${HADOOP_HOME}/etc/hadoop/hadoop-env.sh >> temp.txt
+cat temp.txt > ${HADOOP_HOME}/etc/hadoop/hadoop-env.sh
 
-rm -rf HDFS/
+if [ -d "HDFS" ]; then
+    rm -rf HDFS/
+fi
 
 if [ ! -d "HDFS" ]
 then
@@ -55,60 +60,61 @@ then
     mkdir -p HDFS/nodemanlog
 fi
 
-num=$(cat $HADOOP_PREFIX/etc/hadoop/core-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
-head -n $(($num - 1)) $HADOOP_PREFIX/etc/hadoop/core-site.xml > temp.txt
-cat temp.txt > $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "<configuration>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "        <name>fs.defaultFS</name>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "        <value>hdfs://$master:9000</value>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
-echo "</configuration>" >> $HADOOP_PREFIX/etc/hadoop/core-site.xml
+num=$(cat ${HADOOP_HOME}/etc/hadoop/core-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
+head -n $(($num - 1)) ${HADOOP_HOME}/etc/hadoop/core-site.xml > temp.txt
+cat temp.txt > ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "<configuration>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "        <name>fs.defaultFS</name>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "        <value>hdfs://$master:9000</value>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
+echo "</configuration>" >> ${HADOOP_HOME}/etc/hadoop/core-site.xml
 
-num=$(cat $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
-head -n $(($num - 1)) $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml > temp.txt
-cat temp.txt > $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "<configuration>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <name>dfs.replication</name>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <value>1</value>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <name>dfs.namenode.name.dir</name>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <value>$(pwd)/HDFS/namenode</value>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <name>dfs.datanode.data.dir</name>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "        <value>$(pwd)/HDFS/datanode</value>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
-echo "</configuration>" >> $HADOOP_PREFIX/etc/hadoop/hdfs-site.xml
+num=$(cat ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
+head -n $(($num - 1)) ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml > temp.txt
+cat temp.txt > ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "<configuration>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <name>dfs.replication</name>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <value>1</value>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <name>dfs.namenode.name.dir</name>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <value>$(pwd)/HDFS/namenode</value>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <name>dfs.datanode.data.dir</name>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "        <value>$(pwd)/HDFS/datanode</value>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
+echo "</configuration>" >> ${HADOOP_HOME}/etc/hadoop/hdfs-site.xml
 
-num=$(cat $HADOOP_PREFIX/etc/hadoop/yarn-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
-head -n $(($num - 1)) $HADOOP_PREFIX/etc/hadoop/yarn-site.xml > temp.txt
-cat temp.txt > $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "<configuration>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "        <name>yarn.nodemanager.aux-services</name>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "        <value>mapreduce_shuffle</value>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
-echo "</configuration>" >> $HADOOP_PREFIX/etc/hadoop/yarn-site.xml
+num=$(cat ${HADOOP_HOME}/etc/hadoop/yarn-site.xml | grep -n "<configuration>" | cut -d ':' -f1)
+head -n $(($num - 1)) ${HADOOP_HOME}/etc/hadoop/yarn-site.xml > temp.txt
+cat temp.txt > ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "<configuration>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "        <name>yarn.nodemanager.aux-services</name>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "        <value>mapreduce_shuffle</value>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
+echo "</configuration>" >> ${HADOOP_HOME}/etc/hadoop/yarn-site.xml
 
-num=$(cat $HADOOP_PREFIX/etc/hadoop/mapred-site.xml.template | grep -n "<configuration>" | cut -d ':' -f1)
-head -n $(($num - 1)) $HADOOP_PREFIX/etc/hadoop/mapred-site.xml.template > temp.txt
-cat temp.txt > $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "<configuration>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "        <name>mapreduce.framework.name</name>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "        <value>yarn</value>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "    <property>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "        <name>mapreduce.fileoutputcommitter.marksuccessfuljobs</name>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "        <value>false</value>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "    </property>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
-echo "</configuration>" >> $HADOOP_PREFIX/etc/hadoop/mapred-site.xml
+num=$(cat ${HADOOP_HOME}/etc/hadoop/mapred-site.xml.template | grep -n "<configuration>" | cut -d ':' -f1)
+head -n $(($num - 1)) ${HADOOP_HOME}/etc/hadoop/mapred-site.xml.template > temp.txt
+cat temp.txt > ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "<configuration>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "        <name>mapreduce.framework.name</name>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "        <value>yarn</value>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "    <property>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "        <name>mapreduce.fileoutputcommitter.marksuccessfuljobs</name>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "        <value>false</value>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "    </property>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
+echo "</configuration>" >> ${HADOOP_HOME}/etc/hadoop/mapred-site.xml
 
-echo "export HADOOP_HOME=$HADOOP_PREFIX" > hadoop_prefix.sh
-echo "export HADOOP_CONF_DIR=$HADOOP_PREFIX/etc/hadoop" >> hadoop-prefix.sh
-echo "export YARN_CONF_DIR=$HADOOP_PREFIX/etc/hadoop" >> hadoop-prefix.sh
-echo "export PATH=$HADOOP_PREFIX/bin:$HADOOP_PREFIX/sbin:$PATH" >> hadoop-prefix.sh
+echo "export HADOOP_HOME=$HADOOP_HOME" > hadoop-prefix.sh
+echo "export HADOOP_CONF_DIR=$HADOOP_HOME/etc/hadoop" >> hadoop-prefix.sh
+echo "export LD_LIBRARY_PATH=$HADOOP_HOME/lib/native:$LD_LIBRARY_PATH" >> hadoop-prefix.sh
+echo "export YARN_CONF_DIR=$HADOOP_HOME/etc/hadoop" >> hadoop-prefix.sh
+echo "export PATH=${HADOOP_HOME}/bin:${HADOOP_HOME}/sbin:$PATH" >> hadoop-prefix.sh
